+++ megatron/layers/wrappers.py
@@ -1,7 +1,7 @@
 from .core import Layer
 
 
-class SklearnLayer(Layer):
+class Sklearn(Layer):
     def __init__(self, sklearn_transformation):
         super().__init__()
         self.transformation = sklearn_transformation
@@ -21,4 +21,22 @@ class SklearnLayer(Layer):
         if hasattr(self.transformation, 'transform'):
             return self.transformation.transform(*inputs)
         else:
-            return self.transformation.predict(*inputs)
+            # don't use the labels for this
+            return self.transformation.predict(inputs[0])
+
+
+class Keras(Layer):
+    def __init__(self, keras_model):
+        super().__init__()
+        self.model = keras_model
+        self.name = 'KerasModel'
+
+    def partial_fit(self, *inputs):
+        self.model.fit(*inputs)
+
+    def fit(self, *inputs):
+        self.model.fit(*inputs)
+
+    def transform(self, *inputs):
+        # don't use the labels for this
+        return self.model.predict(inputs[0])